实装自动移动

+++ Assets/MusicGame/Scripts/Model/RhythmPointInfoModel.cs
@@ -7,14 +7,14 @@ namespace AssemblyCSharp.Assets.MusicGame
     {
         public PointType type { get; private set; }
         public Vector3 vector3 { get; private set; }
-        public double time { get; private set; }
+        public float time { get; private set; }
 
         public object Clone()
         {
             return this.MemberwiseClone();
         }
 
-        public void setValues(PointType type, Vector3 vector3, double time)
+        public void setValues(PointType type, Vector3 vector3, float time)
         {
             this.type = type;
             this.vector3 = vector3;
